David's suggestions pre client meeting.
We change the field display ratios to label col-md-7, value col-md-5 to
allow for long field names.

We removed the patient outcome at the last visit as the client says that
this can be inferred from the rest of the data.

We change previous neoplasm to neoplasm as its already in the past
medical history section so the previous could be inferred.

There is no check box for second neoplasm, so we remove that field and
use the neoplasm field for a boolean to mean one or more neoplasm.